refactor: split view.rs into layout, popups, and status modules
